Creating Subfolders for String Topic Trie
