+++ String/Trie/UVA - 11362 Solution With Trie Concept.cpp
@@ -1,0 +1,74 @@
+#pragma warning(disable:4786)
+#pragma warning(disable:4996)
+#include<bits/stdc++.h>
+#define pii pair<int,int>
+#define pll pair<long long ,long long>
+#define pli pair<long long , int>
+#define pil pair<int ,long long>
+#define pi acos(-1)
+#define pb push_back
+#define mkp make_pair
+#define all(a) a.begin(), a.end()
+#define IOS ios::sync_with_stdio(0); cin.tie(0); cout.tie(0);
+#define ll long long
+#define MAX 300005
+#define INF 0x3f3f3f3f
+template <class T> inline T bigmod(T p,T e,T M){ll ret = 1LL;for(; e > 0LL; e >>= 1LL){if(e & 1LL) ret = (ret * p) % M;p = (p * p) % M;}return (T)ret;}
+template <class T> inline T modinverse(T a,T M){return bigmod(a,M-2,M);}   // M is prime}
+using namespace std;
+
+int TRIE[MAX][10];
+int tot;
+bool endnode[MAX];
+
+int insert(string s){
+    int p=1;
+    int found=0;
+    for(int i=0;i<s.length();i++){
+        int temp = s[i]-'0';
+        if(TRIE[p][temp]==0){
+            TRIE[p][temp]=++tot;
+        }
+        p = TRIE[p][temp];
+        if(endnode[p]==1){
+            found=1;
+        }
+    }
+    endnode[p]=1;
+    return found;
+}
+
+bool cmp(string s1, string s2){
+    return s1.length()<s2.length();
+}
+
+int main(){
+    IOS
+    //freopen("output.txt","w",stdout);
+    int t;
+    cin>>t;
+    while(t--){
+        tot=1;
+        memset(TRIE,0,sizeof TRIE);
+        memset(endnode,0,sizeof endnode);
+        int n;
+        cin>>n;
+        string s[n];
+        for(int i=0;i<n;i++){
+            cin>>s[i];
+        }
+        sort(s,s+n,cmp);
+        int found=0;
+
+        for(int i=0;i<n;i++){
+            found = insert(s[i]);
+            if(found)break;
+        }
+        if(found){
+            cout<<"NO\n";
+        }
+        else{
+            cout<<"YES\n";
+        }
+    }
+}
